+++ src/app/meetings/modal/model/meeting-image.types.ts
