+++ src/app/meetings/modal/services/uploadMeetingImage.ts
